Moved pagination and ratelimit tests to their own test files
